create syncmetric package
